src/requirements_dev.txt
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.11
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.0
django-debug-permissions==1.0.0
django-debug-toolbar==2.2
django-durationwidget==1.0.5
django-formset-js==0.5.0
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-mass-edit==3.2.0
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-post-office==3.4.0
django-selectable==1.2.1
django-tenants==3.1.0
django-widget-tweaks==1.4.8
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
phonenumbers==8.12.6
psycopg2==2.8.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
